chore: clean up tcs
